pppColum: use float drawScale in pppRenderColum (#5997)
* pppColum: use float drawScale in pppRenderColum

* pppColum: keep signed compare form in pppFrameColum

---------

+++ src/pppColum.cpp
@@ -76,7 +76,7 @@ void pppRenderColum(pppColum *column, pppColumUnkB *param_2, pppColumUnkC *param
             float lengthXY;
             float fadeRange = *(float*)(param_2->m_payload + 0x10);
             float segmentStep;
-            double drawScale = 0.0;
+            float drawScale = 0.0f;
 
             PSMTXIdentity(identityMtx);
             basePos.x = *(float*)(workPos + 0x10);
@@ -140,8 +140,8 @@ void pppRenderColum(pppColum *column, pppColumUnkB *param_2, pppColumUnkC *param
                     pppGetShapeUV__FPlsR5Vec2dR5Vec2di((long*)shapeSt->m_animData,
                                                        *(short*)(workFrame + 2), uvA, uvB, j);
 
-                    PSVECScale(&shapePosA, &shapePosA, (float)drawScale);
-                    PSVECScale(&shapePosB, &shapePosB, (float)drawScale);
+                    PSVECScale(&shapePosA, &shapePosA, drawScale);
+                    PSVECScale(&shapePosB, &shapePosB, drawScale);
                     PSVECAdd(&shapePosA, &center, &shapePosA);
                     PSVECAdd(&shapePosB, &center, &shapePosB);
